note/8_分笔数据实时订阅与多Client问题/v1.0/windows_端/requirements.txt
# Windows端依赖包
redis==4.5.4
PyYAML==6.0
xtquant>=1.0.0  # QMT量化交易接口
